requirements.txt
comet-ml==3.33.9
dataclasses-json==0.5.7
numpy==1.23.2
pexpect==4.8.0
sentence-transformers==2.2.2
#torch==1.12.1
#transformers==4.21.2
sexpdata==1.0.0
pampy==0.3.0
ray==2.4.0
pydantic==1.10.12 # pydantic > 2 is not compatible with ray
faiss-cpu>=1.6.1
filelock
#numpy
regex
torch>=1.5.0
transformers>=4.3
tqdm>=4.27
wget
spacy>=2.1.8
hydra-core>=1.0.0
omegaconf>=2.0.1
jsonlines
soundfile==0.12.1
editdistance==0.6.2
rank_bm25==0.2.2
parglare==0.16.1
openai==0.27.8
tiktoken==0.4.0
psutil==5.9.5
accelerate==0.20.3
urllib3>=2.0.7
mathlibtools==1.3.2
litellm==0.8.6
huggingface-hub==0.17.3
#protobuf==3.20.1 # Needed for PISA Isabelle
anthropic~=0.30.0
replicate~=0.27.0
google-generativeai